move yaml related utils to their own module
