+++ tests/test_cli/test_config.py
@@ -30,7 +30,7 @@ from aea.cli import cli
 from aea.cli.utils.constants import ALLOWED_PATH_ROOTS
 from aea.configurations.base import AgentConfig, DEFAULT_AEA_CONFIG_FILE, PackageType
 from aea.configurations.loader import ConfigLoader
-from aea.helpers.base import yaml_load
+from aea.helpers.yaml_utils import yaml_load
 
 from tests.conftest import CLI_LOG_OPTION, CUR_PATH, CliRunner
